--- src/translate/languages.js
@@ -1,18 +1,0 @@
-const portuguese = {
-   
-    header:{
-        home: "INICIO",
-        projects: "PROJETOS",
-        repos: "REPOSITORIOS"
-    }
-};
-
-const english = {
-    header:{
-        home: "HOME",
-        projects: "PROJECTS",
-        repos: "REPOSITORIES"
-    }
-};
-
-export { portuguese, english };
